+++ spec/spec_helper.rb
@@ -6,12 +6,14 @@ module ConfigHelpers
     Namecheap.config.username = nil
     Namecheap.config.key = nil
     Namecheap.config.client_ip = nil
+    Namecheap.config.environment = nil
   end
 
   def set_dummy_config
     Namecheap.config.username = "the_username"
     Namecheap.config.key = "the_key"
     Namecheap.config.client_ip = "127.0.0.1"
+    Namecheap.config.environment = nil
   end
 end